File structure formatting done
